applying classification

+++ app.py
@@ -11,7 +11,7 @@ def home():
 def predict():
     int_features = [x for x in request.form.values()]
     final_features = [np.array(int_features)]
-    prediction = model.predict(final_features).astype(int)
+    prediction = model.predict(final_features)
 
     return render_template('index.html', prediction_text='QUALITY OF WINE IS:{}'.format(prediction))
